refactor(brain): extract yaml-parse, confidence, dream-plan, dream-refresh, reconcile-outcomes modules
policy.ts (1871 lines) embedded a 230-line YAML subset parser;
dream.ts (2239 lines) mixed scan-record shapes, plan state,
confidence math, evidence refresh, and contradiction reconcile in one
file. Each concern now lives in its own module with explicit exports:

- yaml-parse.ts - the _brain.yaml subset parser (grammar unchanged).
- dream-plan.ts - scan-record and plan-state shapes plus the two
  stateless plan helpers shared by the planning sub-modules.
- confidence.ts - Wilson-bound confidence computation and banding.
- dream-refresh.ts - apply-evidence scan (merge-alias aware) and the
  refresh/promotion planner, including outcome regression.
- reconcile-outcomes.ts - contradiction classification (phase F3).

dream.ts shrinks to the orchestrator (1499 lines), policy.ts to config
semantics (1596 lines). Importers migrated directly; no re-export
shims. New unit tests cover each extracted module in isolation.

Kanban: t_3d9369d0
