+++ src/core/brain/dream.ts
@@ -41,20 +41,13 @@ import { parseFrontmatter } from "../vault.ts";
 import { regenerateActiveQuiet } from "./active.ts";
 import { openWorkrun, WORKRUN_PHASE, type WorkrunHandle } from "./dream-workrun.ts";
 import { DREAM_PHASE, type DreamPhase, type DreamPhaseSummary } from "./dream-phases.ts";
-import {
-  resolveContradiction,
-  type ContradictionInput,
-  type ReconcileSignal,
-} from "./reconcile-domains.ts";
 import { extractTemporalConstraints } from "./temporal-extract.ts";
 import { runHealEnrichment } from "./heal-run.ts";
 import { collectEvidenceForSlug } from "./evidence.ts";
-import { classifyFreshnessTrend } from "./temporal/freshness-trend.ts";
 import { buildIntentReview, type BrainIntentReviewEntry } from "./intent-review.ts";
 import { writePreferenceTxn } from "./preference-txn.ts";
 import { appendLogEvent, type BrainLogEntry } from "./log.ts";
-import { listLogDates, readLogDay } from "./log-jsonl.ts";
-import { moveToRetired, parsePreference, wouldRewritePreference } from "./preference.ts";
+import { moveToRetired, parsePreference } from "./preference.ts";
 import { parseSignal } from "./signal.ts";
 import { isPinned } from "./pin.ts";
 import { createSnapshot, pruneSnapshots } from "./snapshot.ts";
@@ -70,10 +63,26 @@ import {
 } from "./paths.ts";
 import { countSigns, dominantSignOf } from "./sign.ts";
 import { isoDate, isoSecond } from "./time.ts";
-import { parseWikilink, renderPrefLink } from "./wikilink.ts";
 import {
-  BRAIN_APPLY_RESULT,
-  BRAIN_CONFIDENCE,
+  emptyPlan,
+  filterWithinWindow,
+  recordSignalMove,
+  type CorruptedEntry,
+  type PlanState,
+  type PreferenceRecord,
+  type RetiredRecord,
+  type ScanResult,
+  type SignalRecord,
+} from "./dream-plan.ts";
+import {
+  planRefresh,
+  scanApplyEvidence,
+  type DreamOutcomeRegression,
+  type RefreshResult,
+} from "./dream-refresh.ts";
+import { buildReconcileOutcomes } from "./reconcile-outcomes.ts";
+import { renderPrefLink } from "./wikilink.ts";
+import {
   BRAIN_LOG_EVENT_KIND,
   BRAIN_PREFERENCE_STATUS,
   BRAIN_RETIRED_REASON,
@@ -272,43 +281,6 @@ export interface DreamOptions {
 }
 
 // ----- Internal scan types ------------------------------------------------
-
-interface SignalRecord {
-  readonly path: string;
-  readonly signal: BrainSignal;
-  /** True iff the file lives in `inbox/` (not `processed/`). */
-  readonly active: boolean;
-}
-
-interface PreferenceRecord {
-  readonly path: string;
-  readonly pref: BrainPreference;
-}
-
-interface RetiredRecord {
-  readonly path: string;
-  readonly topic: string;
-  readonly id: string;
-  readonly principle: string;
-  readonly scope?: string;
-  /**
-   * The free-form user reason passed to `o2b brain reject --reason`.
-   * Presence triggers signal-suppression for future signals on the
-   * same (topic, scope) — see §6 of the OSB features summary.
-   */
-  readonly user_rejected_reason?: string;
-}
-
-interface CorruptedEntry {
-  readonly path: string;
-}
-
-interface ScanResult {
-  readonly signals: SignalRecord[];
-  readonly preferences: PreferenceRecord[];
-  readonly retired: RetiredRecord[];
-  readonly corrupted: CorruptedEntry[];
-}
 
 // ----- Main entry ----------------------------------------------------------
 
@@ -997,113 +969,6 @@ function scanBrain(vault: string): ScanResult {
 
 // ----- Planning -----------------------------------------------------------
 
-interface PlanState {
-  /** Topic slug → planned new unconfirmed preference. */
-  readonly newUnconfirmed: NewUnconfirmedPlan[];
-  /** Preferences to retire (after refresh). */
-  readonly retires: RetirePlan[];
-  /** Same-sign signals on active prefs → moved + log event. */
-  readonly notedRedundant: NotedRedundantPlan[];
-  /** Pinned prefs that would have retired but stay because pinned. */
-  readonly retainPinned: RetainPinnedPlan[];
-  /** Signal id → record to move out of inbox/. */
-  readonly signalsToMove: Map<string, SignalMovePlan>;
-  /** Topic slugs flagged contradicted but no transition this run. */
-  readonly contradictionTopics: Set<string>;
-  /**
-   * Signals dropped because their (topic, scope) matches a user-rejected
-   * retired pref carrying a `user_rejected_reason`. Each entry produces
-   * one `signal-suppressed` log event AND a move into `processed/` so
-   * the inbox does not accumulate.
-   */
-  readonly signalsSuppressed: SignalSuppressedPlan[];
-  /**
-   * Signal clusters held back from promotion by the self-approval
-   * guardrail (v0.10.16). The cluster passed the existing
-   * `candidate_threshold` but failed one or more configured
-   * thresholds in `BrainGuardrailConfig`. Preserved across the plan
-   * so it surfaces on the DreamRunSummary without affecting the
-   * existing move-to-processed semantics.
-   */
-  readonly quarantined: DreamQuarantinedEntry[];
-}
-
-interface SignalSuppressedPlan {
-  readonly signal: string;
-  /** Pre-rendered `[[ret-slug|principle]]` wikilink for the suppressor. */
-  readonly retired: string;
-  readonly reason: string;
-  readonly topic: string;
-}
-
-interface NewUnconfirmedPlan {
-  readonly slug: string;
-  readonly topic: string;
-  readonly scope: string | undefined;
-  readonly principle: string;
-  readonly evidencedBy: ReadonlyArray<string>;
-  readonly sign: BrainSignalSign;
-  /**
-   * Wikilink string (`[[ret-<slug>]]` or `[[pref-<slug>]]`) to the
-   * preference this new entry supersedes, if any. Threaded through to
-   * `writePreference` so the resulting frontmatter carries
-   * `supersedes:` for audit-trail continuity across rebuttals.
-   */
-  readonly supersedes?: string;
-  /**
-   * Brain lifecycle suite (F5). Bi-temporal validity derived from the
-   * source signal at plan time (explicit signal fields preferred, else
-   * extracted from the signal's ISO temporal text). Threaded to the
-   * preference writer on promotion.
-   */
-  readonly valid_from?: string;
-  readonly valid_until?: string;
-}
-
-interface RetirePlan {
-  readonly slug: string;
-  /**
-   * Principle of the preference being retired, captured at plan time
-   * so the dream summary log payload can render a titled wikilink
-   * (`[[ret-slug|principle]]`) without re-reading the file after move.
-   */
-  readonly principle: string;
-  readonly reason: BrainRetiredReason;
-  readonly supersededBy?: string;
-}
-
-interface NotedRedundantPlan {
-  /** Pre-rendered `[[pref-id|principle]]` wikilink for the active pref. */
-  readonly preference: string;
-  readonly signal: string;
-}
-
-interface RetainPinnedPlan {
-  /** Pre-rendered `[[pref-id|principle]]` wikilink for the pinned pref. */
-  readonly preference: string;
-  readonly reason: BrainRetiredReason;
-}
-
-interface SignalMovePlan {
-  readonly id: string;
-  readonly date: string;
-  readonly slug: string;
-  readonly path: string;
-}
-
-function emptyPlan(): PlanState {
-  return {
-    newUnconfirmed: [],
-    retires: [],
-    notedRedundant: [],
-    retainPinned: [],
-    signalsToMove: new Map(),
-    contradictionTopics: new Set(),
-    signalsSuppressed: [],
-    quarantined: [],
-  };
-}
-
 function planTopics(
   scan: ScanResult,
   cfg: BrainConfig,
@@ -1449,29 +1314,6 @@ function allocatePreferencePlanSlug(base: string, reserved: Set<string>): string
   return candidate;
 }
 
-function recordSignalMove(plan: PlanState, rec: SignalRecord): void {
-  if (!rec.active) return;
-  const id = rec.signal.id;
-  if (plan.signalsToMove.has(id)) return;
-  // Derive date + slug from the id (`sig-YYYY-MM-DD-<slug>`).
-  const m = /^sig-(\d{4}-\d{2}-\d{2})-(.+)$/.exec(id);
-  if (!m) return;
-  plan.signalsToMove.set(id, {
-    id,
-    date: m[1]!,
-    slug: m[2]!,
-    path: rec.path,
-  });
-}
-
-function filterWithinWindow(sigs: SignalRecord[], windowDays: number, now: Date): SignalRecord[] {
-  const minTime = now.getTime() - windowDays * 24 * 3600 * 1000;
-  return sigs.filter((s) => {
-    const t = Date.parse(s.signal.created_at);
-    return Number.isFinite(t) && t >= minTime;
-  });
-}
-
 /**
  * Brain lifecycle suite (F5). Derive a preference's bi-temporal validity
  * from the source signal: prefer the signal's explicit `valid_from` /
@@ -1498,580 +1340,7 @@ function phaseSummary(phase: DreamPhase, metrics: Record<string, number>): Dream
   return { phase, metrics };
 }
 
-/** Project a scanned signal onto the minimal reconcile view. */
-function toReconcileSignal(r: SignalRecord): ReconcileSignal {
-  return {
-    created_at: r.signal.created_at,
-    ...(r.signal.recorded_at ? { recorded_at: r.signal.recorded_at } : {}),
-    ...(r.signal.source ? { source: r.signal.source } : {}),
-  };
-}
-
-// ----- Reconcile (F3) ------------------------------------------------------
-
-interface ReconcileAutoResolved {
-  readonly topic: string;
-  readonly domain: "source-freshness";
-  readonly winner_sign: "positive" | "negative";
-  readonly margin_days: number;
-}
-
-interface ReconcileOutcomes {
-  readonly openQuestions: DreamOpenQuestion[];
-  readonly autoResolved: ReconcileAutoResolved[];
-}
-
-/**
- * Classify every contradiction topic the plan flagged into a domain and
- * decide its outcome. Source-freshness contradictions where the fresher
- * side leads by at least half the contradiction window auto-resolve
- * (recorded only - never a sub-threshold mutation); the rest become
- * operator-facing open questions. Pure; emits nothing.
- */
-function buildReconcileOutcomes(
-  scan: ScanResult,
-  plan: PlanState,
-  cfg: BrainConfig,
-  now: Date,
-): ReconcileOutcomes {
-  const openQuestions: DreamOpenQuestion[] = [];
-  const autoResolved: ReconcileAutoResolved[] = [];
-  if (plan.contradictionTopics.size === 0) return { openQuestions, autoResolved };
-
-  const windowDays = cfg.dream.contradiction_window_days;
-  // Freshness must be decisive WITHIN the contradiction window, so the
-  // margin is half the window (a gap larger than the window would have
-  // excluded the older signal from the contradiction in the first place).
-  const freshnessMarginDays = Math.max(1, Math.ceil(windowDays / 2));
-
-  const byTopic = new Map<string, SignalRecord[]>();
-  for (const rec of scan.signals) {
-    const arr = byTopic.get(rec.signal.topic);
-    if (arr) arr.push(rec);
-    else byTopic.set(rec.signal.topic, [rec]);
-  }
-
-  for (const topic of plan.contradictionTopics) {
-    const sigs = filterWithinWindow(byTopic.get(topic) ?? [], windowDays, now);
-    const positives = sigs
-      .filter((s) => s.signal.signal === BRAIN_SIGNAL_SIGN.positive)
-      .map(toReconcileSignal);
-    const negatives = sigs
-      .filter((s) => s.signal.signal === BRAIN_SIGNAL_SIGN.negative)
-      .map(toReconcileSignal);
-    const scope = sigs.find((s) => s.signal.scope)?.signal.scope;
-    const input: ContradictionInput = {
-      topic,
-      ...(scope ? { scope } : {}),
-      positives,
-      negatives,
-    };
-    const outcome = resolveContradiction(input, { now, freshnessMarginDays });
-    if (outcome.kind === "auto-resolved") {
-      autoResolved.push({
-        topic,
-        domain: outcome.domain,
-        winner_sign: outcome.winner_sign,
-        margin_days: outcome.margin_days,
-      });
-    } else {
-      openQuestions.push(outcome.question);
-    }
-  }
-  return { openQuestions, autoResolved };
-}
-
 // ----- Refresh + promote --------------------------------------------------
-
-interface RefreshUpdate {
-  readonly slug: string;
-  readonly topic: string;
-  readonly scope?: string;
-  readonly principle: string;
-  readonly created_at: string;
-  readonly unconfirmed_until: string;
-  readonly status: (typeof BRAIN_PREFERENCE_STATUS)[keyof typeof BRAIN_PREFERENCE_STATUS];
-  readonly evidenced_by: ReadonlyArray<string>;
-  readonly confirmed_at: string | null;
-  readonly applied_count: number;
-  readonly violated_count: number;
-  readonly last_evidence_at: string | null;
-  readonly confidence: BrainConfidence;
-  readonly confidence_value: number;
-  readonly pinned: boolean;
-  /**
-   * Freshness trend (t_ee09a6ce), computed at PLAN time so the no-op
-   * pre-flight (`wouldRewritePreference`) renders the same bytes the
-   * write loop will stamp. Computing it only in the write loop would
-   * make every stamped pref look like "needs rewrite" forever. Absent
-   * when the stamp is withheld: a pref that has never carried the
-   * field and is not being rewritten anyway stays byte-identical, so
-   * legacy vaults keep their no-op dream guarantees.
-   */
-  readonly freshness_trend?: string;
-}
-
-/**
- * Pref → recorded band drop within the current dream pass. Captured
- * during refresh so the digest can render a `## Confidence drops`
- * section without re-deriving transitions from the log.
- */
-/**
- * Outcome regression (t_d478df53): minimum applied-with-failure events
- * before a preference is flagged, and the multiplicative confidence
- * penalty a flagged preference receives. The penalty is part of the
- * deterministic confidence derivation, so reruns stay no-ops.
- */
-export const OUTCOME_REGRESSION_MIN_FAILURES = 2;
-export const OUTCOME_REGRESSION_PENALTY = 0.8;
-
-export interface DreamOutcomeRegression {
-  readonly id: string;
-  readonly principle: string;
-  readonly failures: number;
-  readonly successes: number;
-}
-
-export interface RefreshBandDrop {
-  readonly id: string;
-  readonly principle: string;
-  readonly previous: BrainConfidence;
-  readonly next: BrainConfidence;
-  readonly applied: number;
-  readonly violated: number;
-  readonly previous_value: number | null;
-  readonly next_value: number;
-}
-
-interface RefreshResult {
-  /** Slugs transitioning unconfirmed → confirmed in THIS run. */
-  readonly confirmed: Set<string>;
-  /** Slug → full updated frontmatter to write. */
-  readonly updated: Map<string, RefreshUpdate>;
-  /**
-   * Preferences whose `confidence` band dropped (e.g. `high → medium`)
-   * during the refresh phase. Newest-first stable order (insertion
-   * order at construction).
-   */
-  readonly bandDrops: RefreshBandDrop[];
-  /** Preferences flagged by the outcome-regression rule (t_d478df53). */
-  readonly outcomeRegressions: DreamOutcomeRegression[];
-}
-
-interface ApplyEvidenceEntry {
-  readonly pref_slug: string;
-  readonly timestamp: string;
-  readonly result: (typeof BRAIN_APPLY_RESULT)[keyof typeof BRAIN_APPLY_RESULT];
-  /** Downstream outcome when recorded (t_d478df53). */
-  readonly outcome?: "success" | "failure";
-}
-
-function scanApplyEvidence(vault: string): ApplyEvidenceEntry[] {
-  const out: ApplyEvidenceEntry[] = [];
-  const mergeAliases = new Map<string, string>();
-  // Shard-aware (Memory Integrity Suite): merged per-day reads.
-  for (const date of listLogDates(vault)) {
-    const { entries } = readLogDay(vault, date);
-    for (const e of entries) {
-      if (e.eventType === BRAIN_LOG_EVENT_KIND.merge) {
-        const keep = parseWikilinkFromBodyValue(e.body["keep"]);
-        const drop = parseWikilinkFromBodyValue(e.body["drop"]);
-        if (keep?.startsWith("pref-") && drop?.startsWith("pref-")) {
-          mergeAliases.set(
-            drop.slice("pref-".length),
-            resolveMergeAlias(keep.slice("pref-".length), mergeAliases),
-          );
-        }
-        continue;
-      }
-      if (e.eventType !== BRAIN_LOG_EVENT_KIND.applyEvidence) continue;
-      const prefRaw = e.body["preference"];
-      const result = e.body["result"];
-      if (typeof prefRaw !== "string" || typeof result !== "string") continue;
-      if (
-        result !== BRAIN_APPLY_RESULT.applied &&
-        result !== BRAIN_APPLY_RESULT.violated &&
-        result !== BRAIN_APPLY_RESULT.outdated
-      )
-        continue;
-      // Parse `[[pref-slug]]` → slug.
-      const target = parseWikilink(prefRaw);
-      if (!target || !target.startsWith("pref-")) continue;
-      out.push({
-        pref_slug: target.slice("pref-".length),
-        timestamp: e.timestamp,
-        result: result as ApplyEvidenceEntry["result"],
-        ...(e.body["outcome"] === "success" || e.body["outcome"] === "failure"
-          ? { outcome: e.body["outcome"] as "success" | "failure" }
-          : {}),
-      });
-    }
-  }
-  const resolved = out.map((entry) => ({
-    ...entry,
-    pref_slug: resolveMergeAlias(entry.pref_slug, mergeAliases),
-  }));
-  // Stable order: by timestamp ascending. Multiple entries at the same
-  // second keep their parse order (parseLogDay returns insertion order).
-  resolved.sort((a, b) => Date.parse(a.timestamp) - Date.parse(b.timestamp));
-  return resolved;
-}
-
-function parseWikilinkFromBodyValue(value: unknown): string | null {
-  return typeof value === "string" ? parseWikilink(value) : null;
-}
-
-function resolveMergeAlias(slug: string, aliases: ReadonlyMap<string, string>): string {
-  let current = slug;
-  const seen = new Set<string>();
-  while (!seen.has(current)) {
-    seen.add(current);
-    const next = aliases.get(current);
-    if (!next) return current;
-    current = next;
-  }
-  return current;
-}
-
-function planRefresh(
-  vault: string,
-  scan: ScanResult,
-  evidence: ApplyEvidenceEntry[],
-  cfg: BrainConfig,
-  now: Date,
-  plan: PlanState,
-): RefreshResult {
-  const confirmed = new Set<string>();
-  const updated = new Map<string, RefreshUpdate>();
-  const bandDrops: RefreshBandDrop[] = [];
-  const outcomeRegressions: DreamOutcomeRegression[] = [];
-
-  // Index evidence by slug.
-  const bySlug = new Map<string, ApplyEvidenceEntry[]>();
-  for (const e of evidence) {
-    const arr = bySlug.get(e.pref_slug);
-    if (arr) arr.push(e);
-    else bySlug.set(e.pref_slug, [e]);
-  }
-
-  // Slugs that will retire this run — we skip refresh for those so the
-  // retired/ snapshot reflects the pre-refresh counters (which is the
-  // existing test expectation in moveToRetired).
-  const retiringSlugs = new Set(plan.retires.map((r) => r.slug));
-
-  for (const rec of scan.preferences) {
-    const slug = rec.pref.id.startsWith("pref-") ? rec.pref.id.slice("pref-".length) : rec.pref.id;
-    if (retiringSlugs.has(slug)) continue;
-
-    const ev = bySlug.get(slug) ?? [];
-    const applied = ev.filter((e) => e.result === BRAIN_APPLY_RESULT.applied).length;
-    const violated = ev.filter((e) => e.result === BRAIN_APPLY_RESULT.violated).length;
-    const outdatedCount = ev.filter((e) => e.result === BRAIN_APPLY_RESULT.outdated).length;
-    const lastEvidence = ev.length > 0 ? ev[ev.length - 1]!.timestamp : null;
-    const firstApplied = ev.find((e) => e.result === BRAIN_APPLY_RESULT.applied);
-
-    // `outdated` is a context-driven retire signal: a single event
-    // means the rule's scope still matches but the artifact shows
-    // the rule itself is obsolete (framework migration, convention
-    // change). Pin protects against decay-based retires but NOT
-    // against context-driven ones — pinning means "I want this
-    // rule"; an `outdated` event means "context says this rule no
-    // longer applies anywhere." Honour the explicit signal.
-    //
-    // Idempotency: once retired, the pref moves to `retired/` and
-    // future dream passes don't re-process it from `preferences/`.
-    if (outdatedCount > 0) {
-      plan.retires.push({
-        slug,
-        principle: rec.pref.principle,
-        reason: BRAIN_RETIRED_REASON.supersededByContext,
-      });
-      continue;
-    }
-
-    let status = rec.pref.status;
-    let confirmedAt = rec.pref.confirmed_at;
-    if (status === BRAIN_PREFERENCE_STATUS.unconfirmed && firstApplied) {
-      status = BRAIN_PREFERENCE_STATUS.confirmed;
-      confirmedAt = firstApplied.timestamp;
-      confirmed.add(slug);
-    }
-
-    // Quarantine transitions — only applicable to already-confirmed
-    // and already-quarantined preferences. An unconfirmed pref still
-    // promotes via the firstApplied branch above; quarantine entry is
-    // measured against `confirmed` counts. Detailed semantics live on
-    // `BRAIN_PREFERENCE_STATUS.quarantine` in `types.ts`.
-    if (
-      status === BRAIN_PREFERENCE_STATUS.confirmed &&
-      violated >= applied &&
-      applied > cfg.confidence.low_max_applied
-    ) {
-      status = BRAIN_PREFERENCE_STATUS.quarantine;
-    } else if (status === BRAIN_PREFERENCE_STATUS.quarantine) {
-      const newViolated = violated > rec.pref.violated_count;
-      if (newViolated) {
-        if (isPinned(rec.pref)) {
-          plan.retainPinned.push({
-            preference: renderPrefLink({
-              id: rec.pref.id,
-              principle: rec.pref.principle,
-            }),
-            reason: BRAIN_RETIRED_REASON.quarantineViolated,
-          });
-        } else {
-          plan.retires.push({
-            slug,
-            principle: rec.pref.principle,
-            reason: BRAIN_RETIRED_REASON.quarantineViolated,
-          });
-          // Skip refresh — moveToRetired will read the on-disk
-          // counters when it builds the retired snapshot.
-          continue;
-        }
-      } else if (applied > violated) {
-        status = BRAIN_PREFERENCE_STATUS.confirmed;
-      }
-    }
-
-    const baseConfidence = computeConfidence(applied, violated, lastEvidence, cfg, now);
-
-    // Outcome regression (t_d478df53): applied events whose downstream
-    // outcome was `failure` outnumbering `success` mean the rule looks
-    // confirmed while actively hurting. The flag stages an explainable
-    // finding and a multiplicative confidence penalty - retirement
-    // still goes through the existing gates, never silently. The
-    // penalty is a pure function of the evidence, so reruns converge.
-    const appliedFailures = ev.filter(
-      (e) => e.result === BRAIN_APPLY_RESULT.applied && e.outcome === "failure",
-    ).length;
-    const appliedSuccesses = ev.filter(
-      (e) => e.result === BRAIN_APPLY_RESULT.applied && e.outcome === "success",
-    ).length;
-    const regressed =
-      appliedFailures >= OUTCOME_REGRESSION_MIN_FAILURES && appliedFailures > appliedSuccesses;
-    const confidence = regressed
-      ? rebandConfidence(
-          Math.round(baseConfidence.value * OUTCOME_REGRESSION_PENALTY * 10000) / 10000,
-          cfg,
-        )
-      : baseConfidence;
-    if (regressed) {
-      outcomeRegressions.push({
-        id: rec.pref.id,
-        principle: rec.pref.principle,
-        failures: appliedFailures,
-        successes: appliedSuccesses,
-      });
-    }
-
-    // Idempotency on a no-op rerun: skip refresh for prefs where
-    // counters AND status are unchanged AND the on-disk body already
-    // matches what we would render with current evidence. The second
-    // half (body comparison) is what carries the v0.10.1 migration
-    // forward — a pref whose counters are stable but whose body
-    // still has the v0.9.x placeholder will fail the body check and
-    // get rewritten on the next pass.
-    const previousValue = rec.pref.confidence_value;
-    // A `null` previous value (file not yet touched by dream) is
-    // treated as "matches whatever we just computed" — the
-    // body-bytes check in `wouldRewritePreference` is what triggers
-    // the write that populates the field.
-    const valueDifferent =
-      previousValue !== null && Math.abs(previousValue - confidence.value) > 1e-6;
-    const countersChanged =
-      applied !== rec.pref.applied_count ||
-      violated !== rec.pref.violated_count ||
-      lastEvidence !== rec.pref.last_evidence_at ||
-      status !== rec.pref.status ||
-      confirmedAt !== rec.pref.confirmed_at ||
-      confidence.band !== rec.pref.confidence ||
-      valueDifferent;
-
-    // Freshness trend (t_ee09a6ce): classified at PLAN time so the
-    // no-op pre-flight below and the write loop render identical
-    // bytes. The stamp is OPPORTUNISTIC: it lands when the pref is
-    // being rewritten anyway (counters changed) or already carries
-    // the field (then it refreshes on every consolidation pass, the
-    // Hindsight contract). A never-stamped pref with unchanged
-    // counters is left byte-identical so legacy vaults keep their
-    // no-op dream guarantees.
-    const stampTrend = countersChanged || rec.pref.freshness_trend !== undefined;
-    let trendLabel: string | undefined;
-    if (stampTrend) {
-      const trendEv = collectEvidenceForSlug(vault, slug, {
-        sinceIso: rec.pref.created_at,
-        maxApplied: 1000,
-        maxViolated: 1000,
-      });
-      trendLabel = classifyFreshnessTrend({
-        createdAt: rec.pref.created_at,
-        events: [...trendEv.applied, ...trendEv.violated].map((r) => ({
-          at: r.timestamp,
-          result: r.result,
-        })),
-        nowMs: now.getTime(),
-      }).trend;
-    }
-
-    const prospective = {
-      slug,
-      topic: rec.pref.topic,
-      ...(rec.pref.scope ? { scope: rec.pref.scope } : {}),
-      principle: rec.pref.principle,
-      created_at: rec.pref.created_at,
-      unconfirmed_until: rec.pref.unconfirmed_until,
-      status,
-      evidenced_by: rec.pref.evidenced_by,
-      confirmed_at: confirmedAt,
-      applied_count: applied,
-      violated_count: violated,
-      last_evidence_at: lastEvidence,
-      confidence: confidence.band,
-      confidence_value: confidence.value,
-      pinned: rec.pref.pinned,
-      ...(trendLabel !== undefined ? { freshness_trend: trendLabel } : {}),
-    };
-
-    if (!countersChanged) {
-      const ev2 = collectEvidenceForSlug(vault, slug, {
-        sinceIso: rec.pref.created_at,
-      });
-      if (
-        !wouldRewritePreference(vault, {
-          ...prospective,
-          recentApplied: ev2.applied,
-          recentViolated: ev2.violated,
-          // v0.12.0 Brain Integrity Suite: feed the existing
-          // _revision and _content_hash into the pre-flight render
-          // so the byte comparison matches the writePreferenceTxn
-          // no-op semantic. Otherwise a refreshed pref written by a
-          // prior pass (with _revision: N stamped) would look like
-          // "needs rewrite" against a prospective render that omits
-          // those fields, flipping changed=false runs into spurious
-          // updates.
-          // parsePreference defaults absent _revision to 0; treat 0
-          // the same as "absent" for the pre-flight render so the
-          // bytes-compare matches existing legacy starter files that
-          // never had the field on disk.
-          ...(rec.pref.revision !== undefined && rec.pref.revision > 0
-            ? { revision: rec.pref.revision }
-            : {}),
-          ...(rec.pref.content_hash !== undefined ? { content_hash: rec.pref.content_hash } : {}),
-        })
-      ) {
-        // Counters and body both unchanged — true no-op for this pref.
-        continue;
-      }
-    }
-
-    updated.set(slug, prospective);
-
-    // Capture band drops for the digest. A drop is any transition
-    // where the new band ranks lower than the previous (high →
-    // medium, medium → low, high → low). Stable across re-runs as
-    // long as the underlying counters do — the digest only renders
-    // it when a real transition occurred in this pass.
-    if (BAND_RANK[confidence.band] < BAND_RANK[rec.pref.confidence]) {
-      bandDrops.push(
-        Object.freeze({
-          id: rec.pref.id,
-          principle: rec.pref.principle,
-          previous: rec.pref.confidence,
-          next: confidence.band,
-          applied,
-          violated,
-          previous_value: rec.pref.confidence_value,
-          next_value: confidence.value,
-        }),
-      );
-    }
-  }
-
-  return { confirmed, updated, bandDrops, outcomeRegressions };
-}
-
-/**
- * Confidence computation. Returns both the numeric value and the
- * categorical band derived from it.
- *
- * `value = wilson_low(applied, n) * freshness`:
- *
- *   - `wilson_low(applied, n)` where `n = applied + violated` -
- *     a conservative 95% lower bound on the application rate.
- *     `n == 0` yields `0`.
- *   - `freshness` linearly decays from `1.0` at age 0 to `0.0` at
- *     `retire.stale_evidence_days`. `null` last_evidence_at → `0`.
- *
- * Band thresholds (`confidence.medium_min`, `confidence.high_min`)
- * are applied to `value` directly. The numeric value is what the
- * digest's drop tracker compares across runs.
- */
-export interface ConfidenceComputeResult {
-  readonly value: number;
-  readonly band: BrainConfidence;
-}
-
-const BAND_RANK: Readonly<Record<BrainConfidence, number>> = Object.freeze({
-  low: 0,
-  medium: 1,
-  high: 2,
-});
-
-export function computeConfidence(
-  applied: number,
-  violated: number,
-  lastEvidenceAt: string | null,
-  cfg: BrainConfig,
-  now: Date,
-): ConfidenceComputeResult {
-  const n = applied + violated;
-  let wilsonLow = 0;
-  if (n > 0) {
-    const z = 1.96;
-    const z2 = z * z;
-    const pHat = applied / n;
-    const denom = 1 + z2 / n;
-    const centre = (pHat + z2 / (2 * n)) / denom;
-    const margin = (z * Math.sqrt((pHat * (1 - pHat)) / n + z2 / (4 * n * n))) / denom;
-    wilsonLow = Math.max(0, centre - margin);
-  }
-  let freshness = 0;
-  if (lastEvidenceAt) {
-    const ageMs = now.getTime() - Date.parse(lastEvidenceAt);
-    if (Number.isFinite(ageMs)) {
-      const limitMs = cfg.retire.stale_evidence_days * 24 * 3600 * 1000;
-      if (limitMs > 0) {
-        freshness = Math.max(0, Math.min(1, 1 - ageMs / limitMs));
-      }
-    }
-  }
-  const rawValue = wilsonLow * freshness;
-  const value = Math.round(rawValue * 10000) / 10000;
-
-  let band: BrainConfidence;
-  if (value >= cfg.confidence.high_min) {
-    band = BRAIN_CONFIDENCE.high;
-  } else if (value >= cfg.confidence.medium_min) {
-    band = BRAIN_CONFIDENCE.medium;
-  } else {
-    band = BRAIN_CONFIDENCE.low;
-  }
-  return Object.freeze({ value, band });
-}
-
-/** Re-derive the confidence band for an externally adjusted value. */
-function rebandConfidence(value: number, cfg: BrainConfig): ConfidenceComputeResult {
-  let band: BrainConfidence;
-  if (value >= cfg.confidence.high_min) {
-    band = BRAIN_CONFIDENCE.high;
-  } else if (value >= cfg.confidence.medium_min) {
-    band = BRAIN_CONFIDENCE.medium;
-  } else {
-    band = BRAIN_CONFIDENCE.low;
-  }
-  return Object.freeze({ value, band });
-}
 
 // ----- Auto-retires (expired-unconfirmed, stale-no-evidence) --------------
