+++ tests/core/brain/outcome-evidence.test.ts
@@ -14,7 +14,8 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 
 import { appendApplyEvidence } from "../../../src/core/brain/apply-evidence.ts";
-import { dream, OUTCOME_REGRESSION_PENALTY } from "../../../src/core/brain/dream.ts";
+import { dream } from "../../../src/core/brain/dream.ts";
+import { OUTCOME_REGRESSION_PENALTY } from "../../../src/core/brain/dream-refresh.ts";
 import { collectEvidenceForSlug } from "../../../src/core/brain/evidence.ts";
 import { bootstrapBrain } from "../../../src/core/brain/init.ts";
 import { readLogDay } from "../../../src/core/brain/log-jsonl.ts";
