+++ tests/core/brain/policy-guardrails.test.ts
@@ -5,10 +5,10 @@
 
 import { describe, expect, test } from "bun:test";
 
+import { parseBrainYaml } from "../../../src/core/brain/yaml-parse.ts";
 import {
   BRAIN_GUARDRAIL_DEFAULTS,
   BrainConfigError,
-  parseBrainYaml,
   resolveGuardrails,
   validateBrainConfigDetailed,
 } from "../../../src/core/brain/policy.ts";
